Automatic global remove redundancies completed

--HG--
branch : develop

+++ Ship_Game/GenericButton.cs
@@ -1,7 +1,5 @@
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
-using Microsoft.Xna.Framework.Input;
-using System;
 
 namespace Ship_Game
 {
@@ -34,7 +32,7 @@ namespace Ship_Game
 			this.R = R;
 			this.Text = Text;
 			this.Font = Font;
-			TextPos = new Vector2((float)(R.X + R.Width / 2) - Font.MeasureString(Text).X / 2f, (float)(R.Y + R.Height / 2 - Font.LineSpacing / 2));
+			TextPos = new Vector2(R.X + R.Width / 2 - Font.MeasureString(Text).X / 2f, R.Y + R.Height / 2 - Font.LineSpacing / 2);
 		}
 
 		public GenericButton(Vector2 v, string Text, SpriteFont CapitalFont, SpriteFont SmallFont)
@@ -44,8 +42,8 @@ namespace Ship_Game
 			capT = Text[0].ToString();
 			smallT = Text.Remove(0, 1);
 			R = new Rectangle((int)v.X - (int)CapitalFont.MeasureString(capT).X - (int)SmallFont.MeasureString(Text).X, (int)v.Y, (int)CapitalFont.MeasureString(capT).X + (int)SmallFont.MeasureString(Text).X, CapitalFont.LineSpacing);
-			CapitalPos = new Vector2((float)R.X, (float)R.Y);
-			TextPos = new Vector2(CapitalPos.X + CapitalFont.MeasureString(capT).X + 1f, CapitalPos.Y + (float)CapitalFont.LineSpacing - (float)SmallFont.LineSpacing - 3f);
+			CapitalPos = new Vector2(R.X, R.Y);
+			TextPos = new Vector2(CapitalPos.X + CapitalFont.MeasureString(capT).X + 1f, CapitalPos.Y + CapitalFont.LineSpacing - SmallFont.LineSpacing - 3f);
 		}
 
 		public void Draw(ScreenManager ScreenManager)
@@ -136,13 +134,13 @@ namespace Ship_Game
 
 		public void Transition(Rectangle R)
 		{
-			TextPos = new Vector2((float)(R.X + R.Width / 2) - Font.MeasureString(Text).X / 2f, (float)(R.Y + R.Height / 2 - Font.LineSpacing / 2));
+			TextPos = new Vector2(R.X + R.Width / 2 - Font.MeasureString(Text).X / 2f, R.Y + R.Height / 2 - Font.LineSpacing / 2);
 		}
 
 		public void TransitionCaps(Rectangle R)
 		{
-			CapitalPos = new Vector2((float)R.X, (float)R.Y);
-			TextPos = new Vector2(CapitalPos.X + Cap.MeasureString(capT).X + 1f, CapitalPos.Y + (float)Cap.LineSpacing - (float)Small.LineSpacing - 3f);
+			CapitalPos = new Vector2(R.X, R.Y);
+			TextPos = new Vector2(CapitalPos.X + Cap.MeasureString(capT).X + 1f, CapitalPos.Y + Cap.LineSpacing - Small.LineSpacing - 3f);
 		}
 	}
 }
